% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto_internal.R
\name{.associate_vr_parameters}
\alias{.associate_vr_parameters}
\title{Associates parameters in \code{ParameterValues} with correct names
in \code{VitalRateExpr}}
\usage{
.associate_vr_parameters(db, parsed_kernels)
}
\arguments{
\item{db}{A database object}

\item{parsed_kernels}{the output from \code{.kernel_vr_LHS_RHS_mat}}
}
\value{
A list of lists of quosures containing the expressions and parameters
needed to create a \code{kernel}.
}
\description{
Associates parameters in \code{ParameterValues} with correct names
in \code{VitalRateExpr}
}
