% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto_internal.R
\name{.parse_K_kernel}
\alias{.parse_K_kernel}
\title{parse a K kernel to a quosure}
\usage{
.parse_K_kernel(db)
}
\arguments{
\item{db}{a database object or user created pre-\code{protoIPM}-ish object}
}
\value{
A list with 2 components: kernel components in the form of quosures and
a character vector indicating which sub-kernels are included in the full model
}
\description{
parse a K kernel to a quosure
}
