% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto_internal.R
\name{.split_vr_expr}
\alias{.split_vr_expr}
\title{splits a vital rate expression and returns its components}
\usage{
.split_vr_expr(text)
}
\arguments{
\item{text}{a string with the vital rate expression}
}
\value{
A character vector of parameters
}
\description{
splits a vital rate expression and returns its components
}
