% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{pdb_download}
\alias{pdb_download}
\title{Download PADRINO \code{pdb} objects}
\usage{
pdb_download(save = TRUE, destination = NULL)
}
\arguments{
\item{save}{Write the PDB object to a folder of csv files?}

\item{destination}{Where to write the \code{pdb} object to.}
}
\description{
Download PADRINO from Github.
}
\details{
This does not currently support versioning because there is only
one version. \code{destination} should be a folder name. When \code{save = TRUE},
a set of 12 csv files will be saved in the \code{destination} folder.
}
