% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ipm.R
\name{discrete_extrema}
\alias{discrete_extrema}
\title{Correct for eviction}
\usage{
discrete_extrema(mat)
}
\arguments{
\item{mat}{The growth matrix in need of correction}
}
\value{
A matrix of the same dimension as \code{mat}, but with columns summing
to 1.
}
\description{
Correct for eviction in an IPM matrix using either
\code{discrete_extrema} or \code{truncated_distribution}. The corrections
differ in subtle ways. See details for more information
}
\details{
\code{discrete_extrema} creates discrete stages at either end of the
domain of the state variable in question. It does this by adding all
transition probabilities not included in the matrix into the top row (for cells
corresponding to the lower half of the domain) or bottom row (for cells
corresponding to the upper half of the domain).

\code{truncated_distribution} works differently. This generates a truncated
probability density function using the \code{\link{truncdist}} package and is
called when creating the expressions for the probability density function
describing growth.

Note the differences in when these functions are called. This may affect construction
of custom IPMs from raw data or altering existing IPMs downloaded from the
\code{Padrino} database. Additionally, anecdotal reports suggest that this
doesnt \emph{completely} correct for eviction as remainders of
\code{1 - colSums(mat)} are occasionally between \code{1e-4} and \code{1e-6}
on the cells near the border of the function domain.
}
