% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{pdb_download}
\alias{pdb_download}
\alias{pdb_save}
\alias{pdb_load}
\title{Download PADRINO \code{pdb} objects}
\usage{
pdb_download(save = TRUE, destination = NULL)

pdb_save(pdb, destination = NULL)

pdb_load(path)
}
\arguments{
\item{save}{Write the PDB object to a folder of text files?}

\item{destination}{Where to write the \code{pdb} object to.}

\item{pdb}{A \code{pdb} object.}

\item{path}{The directory where the PADRINO tables are stored}
}
\value{
\code{pdb_download} and \code{pdb_load} return \code{pdb} objects.
\code{pdb_save} returns a \code{pdb} object invisibly.
}
\description{
Download PADRINO from Github.
}
\details{
This does not currently support versioning because there is only
one version. \code{destination} should be a folder name. When \code{save = TRUE},
a set of 12 text files will be saved in the \code{destination} folder. The files
are tab-delimited.
}
