% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_K.R
\name{make_k}
\alias{make_k}
\alias{make_k.ipm_kernels}
\alias{make_k.ipm_system}
\alias{.make_k_impl}
\title{make_k}
\usage{
make_k(ipm, ...)

\method{make_k}{ipm_kernels}(ipm, proto_ipm, ...)

\method{make_k}{ipm_system}(ipm, proto_ipm, ...)

.make_k_impl(ipm, proto_ipm, domains = NULL, domain_lower = NULL,
  domain_upper = NULL, mesh_points = NULL)
}
\arguments{
\item{ipm}{An object of class \code{ipm_system} or \code{ipm_kernels}.}

\item{...}{additional arguments ADD MORE AS NEEDED}

\item{proto_ipm}{an object of class \code{proto_ipm}. This should be the same
\code{proto_ipm} used to make the \code{ipm}.}
}
\value{
For each distinct ipm, a list of length 2: \code{K} is the iteration
kernel, and \code{sub_kernels} are the components of it. If a model has multiple
Ks, then they will be named according to the hierarchical effects that create
them (e.g. for a model with a year effect, K_2000, K_2001, K_2002, etc). Each
part of the list will have the two slots described above.
}
\description{
make_k
}
