% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto.R
\name{make_proto_ipm}
\alias{make_proto_ipm}
\title{Generate proto_ipms from Padrino objects}
\usage{
make_proto_ipm(pdb, ipm_id = NULL)
}
\arguments{
\item{pdb}{A \code{pdb} object}

\item{ipm_id}{Optionally, one or more \code{ipm_id}'s to build. If empty,
all models contained in the \code{pdb} object will be processed into
\code{proto_ipm}'s.}
}
\value{
A list containing one or more \code{proto_ipms}. Names of the list
will correspond to \code{ipm_id}s.
}
\description{
These functions generate \code{proto_ipm} objects from
Padrino Database tables.
}
\details{
\code{proto_ipm} objects contain all of the information needed
to implement an IPM, but stop short of actually generating kernels.
}
