% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto.R
\name{mat_mult}
\alias{mat_mult}
\title{mat_mult}
\usage{
mat_mult(x, y)
}
\arguments{
\item{x, y}{Numeric matrices or vectors}
}
\value{
A numeric matrix or vector
}
\description{
Multiply two matrices
}
\details{
This is a wrapper around \code{\%*\%} that allows Padrino's internal
syntax to be slightly less R-specific. You're probably better off using
\code{\%*\%} for interactive use.
}
