% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_padrino.R
\name{padrino_filter}
\alias{padrino_filter}
\title{Subset a Padrino database object}
\usage{
padrino_filter(db, ...)
}
\arguments{
\item{db}{The database object to subset}

\item{...}{Logical conditions to use for subsetting}
}
\value{
An object of class \code{PadrinoDB}
}
\description{
currently only works by passing \code{ipm_id}. Will extend later
to use any variable.
}
\author{
Sam Levin
}
