% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{pdb}
\alias{pdb}
\title{Selected models from the Padrino Database}
\format{
A list of data frames, each corresponding to a different table in the
Padrino Database.
}
\usage{
pdb
}
\description{
Selected models from the Padrino Database
}
\keyword{datasets}
