% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pdb}
\alias{print.pdb}
\title{Print a \code{pdb} object.}
\usage{
\method{print}{pdb}(x, ...)
}
\arguments{
\item{x}{A \code{pdb} object.}

\item{...}{ignored}
}
\description{
Print a \code{pdb} object.
}
