% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ipm.R, R/parameter_associations.R
\name{.build_kernel_env}
\alias{.build_kernel_env}
\alias{.generate_domain_env}
\alias{.kernel_vr_LHS_RHS_mat}
\title{kernel vital rate expressions}
\usage{
.build_kernel_env(proto_ipm, kernel, domain_env)

.generate_domain_env(domain)

.kernel_vr_LHS_RHS_mat(db, parsed_K)
}
\arguments{
\item{proto_ipm}{An object of class \code{proto_ipm}. This can come from either
the \code{Padrino} database or from a user specified set of functions supplied
in \code{IPMr}.}

\item{db}{a database object or user created pre-\code{protoIPM}-ish object}

\item{parsed_K}{The output from \code{.parse_K}}
}
\description{
Create LHS-RHS matrix of vital rate expressions for all sub-kernels
of a model
}
