% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lambda.pdb_ipm}
\alias{lambda.pdb_ipm}
\alias{right_ev.pdb_ipm}
\alias{left_ev.pdb_ipm}
\title{Padrino methods for \code{ipmr} generics}
\usage{
\method{lambda}{pdb_ipm}(ipm, ...)

\method{right_ev}{pdb_ipm}(ipm, iterations = 100, tolerance = 1e-10, ...)

\method{left_ev}{pdb_ipm}(ipm, iterations = 100, tolerance = 1e-10, ...)
}
\arguments{
\item{ipm}{A \code{pdb_ipm}.}

\item{...}{further arguments passed to \code{\link[ipmr]{lambda}}. Unused for
\code{left_ev} and \code{right_ev} - only present for compatibility.}

\item{iterations}{The number of times to iterate the model to reach
convergence. Default is 100.}

\item{tolerance}{Tolerance to evaluate convergence to asymptotic dynamics.}
}
\description{
Provides wrappers around \code{ipmr} generic functions to extract
some quantities of interest from \code{pdb_ipm}s.
}
