% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipm.R
\name{make_ipm}
\alias{make_ipm}
\title{Construct an IPM from the proto_IPM object}
\usage{
make_ipm(proto_ipm, domains = NULL, mesh_points = NULL,
  domain_lower = NULL, domain_upper = NULL,
  eviction = c("discrete_extremes", "truncated_distributions"),
  truncated_bounds = NULL, quad_routine = c("midpoint", "trapezoid",
  "Gauss-Legendre"), return_all = TRUE)
}
\arguments{
\item{proto_ipm}{An object of class \code{proto_ipm}. This can come from either
the \code{Padrino} database or from a user specified set of functions supplied
in \code{IPMr}.}

\item{domains}{An alternative domain or set of domains to evaluate each function
in the \code{proto_ipm} object. If \code{NULL} (default), then uses the information
in the \code{proto_ipm} object to construct the IPM}

\item{mesh_points}{An integer specifying number of meshpoints to use for integration.}

\item{domain_lower}{The smallest value on the specified domain(s). If using
more than 1 domain, then specify a numeric vector the same length as \code{domains}.}

\item{domain_upper}{The largest value on the specified domain(s). If using
more than 1 domain, then specify a numeric vector the same length as \code{domains}.}

\item{eviction}{The method to correct for eviction. \code{discrete_extremes}
puts all evicted individuals back into the cells on the matrix edge.
\code{truncated_distributions} creates truncated probability density functions
for each one called in the IPM creation.}

\item{truncated_bounds}{The upper and lower bounds for truncation. If \code{NULL}
(default), uses the upper and lower bounds for the domains associated with
the probability density function.}

\item{quad_routine}{The integration type to use. Currently, \code{midpoint} is
the default. Alternatives are \code{trapezoid} and \code{Gauss-Legendre}. Others
may be added later.}

\item{return_all}{Logical indicating whether to return the complete set of
expressions and environments or just the results of kernel evaluation (e.g.
the iteration matrices). Default is \code{TRUE}, but could cause large models
to get clunky. Set to \code{FALSE} if this is happening.}
}
\value{
An object of class \code{ipm_system}.
}
\description{
Construct an IPM from the proto_IPM object
}
