% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_K.R
\name{make_K}
\alias{make_K}
\title{Make the complete IPM kernel (K)}
\usage{
make_K(ipm)
}
\arguments{
\item{A}{\code{padrino_ipm} object.}
}
\value{
A \code{padrino_k} object. This is a list of length > 1. See details
for more information
}
\description{
Generates the \code{K} kernel for a given \code{padrino_ipm}. This
is only applicable to IPMs that have a top-level kernel. Calling this for an
object will just return the original \code{padrino_ipm}.
}
\details{
Populations with multiple discrete classes can be represented as a
series of kernels that move them within and between levels of these classes.
For example, and age x size IPM is more efficiently implemented as a list
corresponding the age specific size kernels as opposed to a megamatrix, which
is mostly filled with memory consuming zeroes. Common metrics can be calculated
via iteration of the matrices using \code{iterate_ipm}.
}
