% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto.R
\name{make_proto_ipm}
\alias{make_proto_ipm}
\title{generate a \code{proto_ipm} object for a given IPM from the \code{Padrino}
database}
\usage{
make_proto_ipm(db, ipms = NULL)
}
\arguments{
\item{db}{an object downloaded from the \code{Padrino} database}

\item{ipms}{The IPM IDs to build \code{proto_ipm} objects for}
}
\value{
A \code{tibble} of \code{proto_ipm} objects. Each IPM sub-kernel will
 have its own row in the tibble
}
\description{
generate a \code{proto_ipm} object for a given IPM from the \code{Padrino}
database
}
