% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proto_clean.R
\name{make_proto_ipm}
\alias{make_proto_ipm}
\title{generate a \code{proto_ipm} object for a given IPM from the \code{Padrino}
database}
\usage{
make_proto_ipm(db, ...)
}
\arguments{
\item{db}{an object downloaded from the \code{Padrino} database}

\item{condition}{logical conditions to subset the database. These are passed to
\code{padrino_filter}. Useful for only building a subset of downloaded IPMs.}
}
\value{
A \code{tibble} of \code{proto_ipm} objects. Each IPM sub-kernel will
 have its own row in the tibble
}
\description{
generate a \code{proto_ipm} object for a given IPM from the \code{Padrino}
database
}
