% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vital_rate_exprs.pdb_proto_ipm_list}
\alias{vital_rate_exprs.pdb_proto_ipm_list}
\alias{vital_rate_exprs.pdb_ipm}
\alias{kernel_formulae.pdb_proto_ipm_list}
\alias{kernel_formulae.pdb_ipm}
\alias{domains.pdb_proto_ipm_list}
\alias{domains.pdb_ipm}
\alias{parameters.pdb_proto_ipm_list}
\alias{parameters.pdb_ipm}
\alias{parameters<-.pdb_proto_ipm_list}
\alias{pop_state.pdb_proto_ipm_list}
\alias{pop_state.pdb_ipm}
\alias{vital_rate_funs.pdb_ipm}
\alias{int_mesh.pdb_ipm}
\title{Accessor functions for (semi) built Padrino objects}
\usage{
\method{vital_rate_exprs}{pdb_proto_ipm_list}(object)

\method{vital_rate_exprs}{pdb_ipm}(object)

\method{kernel_formulae}{pdb_proto_ipm_list}(object)

\method{kernel_formulae}{pdb_ipm}(object)

\method{domains}{pdb_proto_ipm_list}(object)

\method{domains}{pdb_ipm}(object)

\method{parameters}{pdb_proto_ipm_list}(object)

\method{parameters}{pdb_ipm}(object)

\method{parameters}{pdb_proto_ipm_list}(object, ..., ipm_id = NULL) <- value

\method{pop_state}{pdb_proto_ipm_list}(object)

\method{pop_state}{pdb_ipm}(object)

\method{vital_rate_funs}{pdb_ipm}(ipm)

\method{int_mesh}{pdb_ipm}(ipm, full_mesh = TRUE)
}
\arguments{
\item{object}{An object produced by \code{pdb_make_proto_ipm} or
\code{pdb_make_ipm}.}

\item{...}{Ignored}

\item{ipm_id}{The \code{ipm_id} to set the new parameter values for}

\item{value}{The value to insert. Should be a named list where the names
correspond to parameter names and the entries are new parameter values.}
}
\description{
Padrino versions of \code{\link[ipmr]{accessors}}.
}
