% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vital_rate_exprs.pdb_proto_ipm}
\alias{vital_rate_exprs.pdb_proto_ipm}
\alias{vital_rate_exprs.pdb_ipm}
\alias{kernel_formulae.pdb_proto_ipm}
\alias{kernel_formulae.pdb_ipm}
\alias{domains.pdb_proto_ipm}
\alias{domains.pdb_ipm}
\alias{parameters.pdb_proto_ipm}
\alias{parameters.pdb_ipm}
\alias{parameters<-.pdb_proto_ipm}
\alias{parameters<-.pdb_ipm}
\alias{pop_state.pdb_proto_ipm}
\alias{pop_state.pdb_ipm}
\title{Accessor functions for (semi) built Padrino objects}
\usage{
\method{vital_rate_exprs}{pdb_proto_ipm}(object)

\method{vital_rate_exprs}{pdb_ipm}(object)

\method{kernel_formulae}{pdb_proto_ipm}(object)

\method{kernel_formulae}{pdb_ipm}(object)

\method{domains}{pdb_proto_ipm}(object)

\method{domains}{pdb_ipm}(object)

\method{parameters}{pdb_proto_ipm}(object)

\method{parameters}{pdb_ipm}(object)

\method{parameters}{pdb_proto_ipm}(object) <- value

\method{parameters}{pdb_ipm}(object) <- value

\method{pop_state}{pdb_proto_ipm}(object)

\method{pop_state}{pdb_ipm}(object)
}
\arguments{
\item{object}{An object produced by \code{pdb_make_proto_ipm} or
\code{pdb_make_ipm}.}

\item{value}{The value to insert. Should be a named list where the names
correspond to parameter names and the entries are new parameter values.}
}
\description{
Padrino versions of \code{\link[ipmr]{accessors}}.
}
