% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pdb}
\alias{print.pdb}
\alias{print.pdb_proto_ipm_list}
\title{Print a \code{pdb} object.}
\usage{
\method{print}{pdb}(x, ...)

\method{print}{pdb_proto_ipm_list}(x, ...)
}
\arguments{
\item{x}{A \code{pdb} object.}

\item{...}{Only used by \code{pdb_new_fun_form}, otherwise ignored. See details
and examples for usage in \code{pdb_new_fun_form}.}
}
\value{
\code{x} invisibly.
}
\description{
Print a \code{pdb} object.
}
